apiMotoGP/apiMotoGP/requirements.txt
beautifulsoup4==4.6.3
certifi==2018.8.24
chardet==3.0.4
dataclasses==0.6
dj-database-url==0.5.0
Django==2.1.2
django-filter==2.0.0
django-heroku==0.3.1
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.8.2
djongo==1.2.32
dnspython==1.16.0
html5lib==1.0.1
idna==2.7
lxml==4.2.5
Markdown==3.0.1
mongoengine==0.15.3
psycopg2==2.7.5
pymongo==3.7.2
python-dateutil==2.8.0
pytz==2018.5
requests==2.19.1
six==1.11.0
sqlparse==0.2.4
urllib3==1.23
webencodings==0.5.1
whitenoise==4.1.2
